add : 좌석 예약

+++ src/main/java/com/example/studyroom/repository/SeatRepository.java
@@ -3,5 +3,5 @@ package com.example.studyroom.repository;
 import com.example.studyroom.dao.Seat;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SeatRepository extends JpaRepository<Seat, Long> {
+public interface SeatRepository extends JpaRepository<Seat, Integer> {
 }
